move suggestions to own package. Move copy-rpi.sh to provisioning
